Create exercice3
Write a Python function that checks whether a passed string is palindrome or not. A palindrome is a word, phrase, or sequence that reads the same backward as forward, e.g., madam or nurses run.

Write a Python function that takes a number as a parameter and checks if the number is prime or not. A prime number (or a prime) is a natural number greater than 1 and that has no positive divisors other than 1 and itself.

Write a Python function to check whether a number is in a given range.

Write a Python function to calculate the factorial of a number (a non-negative integer). The function accepts the number as an argument.

Write a Python program to reverse a string.

Write a Python function to sum all the numbers in a list.

Write a Python function to find the Max of three numbers.